Move GitHub related functions to a new module
The functions related to GitHub used in the `cli.version.mike.info`
module are moved to a new module `github` to they can be reused.
